+++ Source/DafnyCore/Options/DafnyProject.cs
@@ -63,8 +63,7 @@ public class DafnyProject : IEquatable<DafnyProject> {
     var diskRoot = Path.GetPathRoot(Uri.LocalPath);
     var result = matcher.Execute(fileSystem.GetDirectoryInfoBase(diskRoot));
     var files = result.Files.Select(f => Path.Combine(diskRoot, f.Path));
-    var rootSourceUris = files.Select(file => new Uri(Path.GetFullPath(file))).ToList();
-    return rootSourceUris;
+    return files.Select(file => new Uri(Path.GetFullPath(file))).ToList();
   }
 
   public bool ContainsSourceFile(Uri uri) {
